Integrate Utils with Location
This commit refactors the utility methods to integrate portions into
the existing Location class. This removes the need to include and call
extra methods and better fits the functionality within Locations api

This commit also reorders locations to sort methods in a more
consistent format (along with alphabetically). The same is done for
utils.dart.