Refactor TransitionManager
This commit updates TransitionManager to remove the ReliqueshPrompt
callback since this was moved to the History api. A warning message is
added to notify users when a null confirmation is given in blocking
mode.

Further, this commit changes the definition of Confirmation to take only
a String instead of String or Prompt. The usage of Confirmation always
converts a prompt to a String first, so there is no need for downstream
implementers of Confirmation to worry about converting from Prompt to
String. Utility methods are updated to reflect this, along with relevant
documentation comments.

Finally, this commit adds a validatePath utility method to properly
check the type of an incoming path during History.push and
History.replace.