Simplify Transition Manager
This commit refactors the TransitionManager class to not extend from
Stream. This is not necessary and does not make sense for such a class.
Instead, the TransitionManager has a stream property that should be used
when a consumer wants to listen to changes.

This refactor considerably simplifies the Browser and Hash variants of
TransitionManager, since StreamController provides the necessary
hooks into the stream listener lifecycle state for proper window
event listening.

Lastly, errors due to the refactors are fixed in the History
implementations.