Fix Confirmation Bug in Hash History
This commit fixes a bug in HashHistory due to an improper way of handling
futures with the window event listeners. When a confirmation returned
false, causing a pop action to be reverted, HashHistory would attempt to
wait for an invalid future to complete.

This is fixed by providing the proper state management for handling this
reversion of popping.